kookvid/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aiosignal==1.4.0
asgiref==3.9.1
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.2
colorama==0.4.6
decorator==5.2.1
dj-database-url==3.0.1
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
ffmpeg-python==0.2.0
frozenlist==1.7.0
future==1.0.0
git-filter-repo==2.47.0
gunicorn==23.0.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
moviepy==2.2.1
multidict==6.6.3
numpy==2.3.2
packaging==25.0
pillow==10.4.0
proglog==0.1.12
propcache==0.3.2
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.1.1
requests==2.32.4
sqlparse==0.5.3
tqdm==4.67.1
twilio==9.7.0
tzdata==2025.2
urllib3==2.5.0
yarl==1.20.1
